Move organization functions to GraphHandler
